[mod]修改儲櫃格式

+++ app/src/main/java/com/lhr/water/util/adapter/WaitDealMaterialAdapter.kt
@@ -71,14 +71,14 @@ class WaitDealMaterialAdapter(
                 itemDetail.receivedQuantity!! - viewModel.formRepository.getMaterialQuantityByTempWaitInputGoods(
                     form.reportTitle!!,
                     form.formNumber!!,
-                    itemDetail.materialNumber!!.toInt()
+                    itemDetail.materialNumber!!
                 )
             } else {
 
                 itemDetail.actualQuantity!! - viewModel.formRepository.getMaterialQuantityByTempWaitOutputGoods(
                     form.reportTitle!!,
                     form.formNumber!!,
-                    itemDetail.materialNumber!!.toInt()
+                    itemDetail.materialNumber!!
                 )
             }
